+++ shewedpolo/Path.cpp
@@ -3,4 +3,4 @@
 
 Path::Path() : distance(INT64_MAX), to(-1) {}
 
-Path::Path(float distance, int to) : distance(distance), to(to) {}
+Path::Path(double distance, int to) : distance(distance), to(to) {}
